+++ django/applications/catmaid/models.py
@@ -43,6 +43,7 @@ class UserRole(object):
     Admin = 'Admin'
     Annotate = 'Annotate'
     Browse = 'Browse'
+    Fork = 'Fork'
     Import = 'Import'
     QueueComputeTask = 'QueueComputeTask'
 
@@ -69,6 +70,7 @@ class Project(models.Model):
             ("can_import", "Can import into projects"),
             ("can_queue_compute_task", "Can queue resource-intensive tasks"),
             ("can_annotate_with_token", "Can annotate project using API token"),
+            ("can_fork", "Can create personal copies of projects (only stacks)"),
         )
 
     def __str__(self):