+++ django/applications/catmaid/urls.py
@@ -102,6 +102,7 @@ urlpatterns += [
     url(r'^projects/$', project.projects),
     url(r'^projects/export$', project.export_projects),
     url(r'^(?P<project_id>\d+)/interpolatable-sections/$', project.interpolatable_sections),
+    url(r'^(?P<project_id>\d+)/fork$', project.fork),
 ]
 
 # General stack model access